requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
allennlp==0.9.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.csv==1.0.7
base58==2.1.1
beautifulsoup4==4.11.1
bleach==4.1.0
blis==0.7.8
boto3==1.23.10
botocore==1.26.10
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.7
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.6.1
clang==5.0
click==8.0.4
cloudpickle==2.1.0
conllu==1.3.1
contextvars==2.4
cryptography==37.0.3
cycler==0.11.0
cymem==1.31.2
dataclasses==0.8
datasets==1.18.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
docutils==0.18.1
editdistance==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
entrypoints==0.4
fairscale==0.4.0
feedparser==6.0.10
filelock==3.0.12
flaky==3.7.0
Flask==2.0.3
Flask-Cors==3.0.10
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
gast==0.4.0
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.20
google-api-core==2.8.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.1
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
h5py==3.1.0
homoglyphs==2.0.4
huggingface-hub==0.4.0
idna==3.3
idna-ssl==1.1.0
imagesize==1.4.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
iso-639==0.4.5
itsdangerous==2.0.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.context==4.1.1
jaraco.functools==3.4.0
jaraco.text==3.7.0
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonnet==0.18.0
jsonpickle==2.2.0
jsonschema==4.0.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
lmdb==1.3.0
lxml==4.9.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
more-itertools==8.13.0
multidict==5.2.0
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==0.26.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
nltk==3.6.7
notebook==6.4.10
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.2.0
opt-einsum==3.3.0
ordereddict==1.1
overrides==3.1.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parsimonious==0.9.0
parso==0.7.1
pathlib==1.0.1
pathtools==0.1.2
pathy==0.6.2
patternfork-nosql==3.6
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plac==0.9.6
pluggy==1.0.0
portend==3.0.0
preshed==0.46.4
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyenchant==3.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.2
python-docx==0.8.11
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.17.0
rsa==4.8
s3transfer==0.5.2
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.6.0
setproctitle==1.2.3
sgmllib3k==1.0.0
shortuuid==1.0.9
six==1.15.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
spacy==1.2.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sputnik==0.9.3
sqlitedict==2.0.0
sqlparse==0.4.2
srsly==2.4.3
tempora==4.1.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.13.0
testpath==0.6.0
thinc==5.0.8
threadpoolctl==3.1.0
tokenizers==0.10.3
tomli==1.2.3
torch==1.2.0
torchvision==0.10.1
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==4.5.1
typer==0.4.1
typing-extensions==3.10.0.2
ujson==4.3.0
Unidecode==1.3.4
urllib3==1.26.9
wandb==0.12.20
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.1
word2number==1.1
wrapt==1.12.1
xxhash==3.0.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
